Starting store-to-map plumbing
