+++ src/components/ItineraryLayer/ItineraryLayer.jsx
@@ -1,0 +1,7 @@
+export const ItineraryLayer = props => {
+
+  return (
+    <div></div>
+  )
+
+}
